requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyYAML==6.0.1